fastapi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
aio-pika==9.4.1
pydantic==2.8.2
psycopg2-binary
sqlalchemy
pandas
numpy
pandas-ta
scikit-learn
xgboost
matplotlib